Fix: Preserve lossless ordinal emission.
- Return the parsed segment ordinal as a fixed-width 64-bit value so the
  accepted 18-digit grammar can never narrow to the host word size.
- Drop the saturating accumulator and its ceiling, which collapsed every
  ordinal above the 32-bit maximum onto one value.
- Carry the full ordinal width through the numeric sort key and the
  gathered ordering key it feeds.
- Add boundary regressions around the 32-bit maximum and the widest
  accepted ordinal for parsing, sorting, anchoring, and flat spelling.
- Register the new emit regressions through a scoped test fragment.
- Refresh the size-ceiling register for the two files whose counts moved.

+++ lib/core/include/nucleus/config.h
@@ -19,6 +19,7 @@
 #include <string>
 #include <vector>
 #include <cstddef>
+#include <cstdint>
 #include <utility>
 #include <optional>
 #include <algorithm>
@@ -165,7 +166,7 @@ private:
         std::size_t count;
     };
 
-    using ordinal_key = std::vector<std::pair<std::string, std::size_t>>;
+    using ordinal_key = std::vector<std::pair<std::string, std::uint64_t>>;
 
     static bool gather_path_matches(std::string_view candidate,
                                     std::string_view query)
